reworked structures a bit
