+++ plugins/offload_rendering_client/CMakeLists.txt
@@ -22,6 +22,7 @@ add_library(${PLUGIN_NAME} SHARED plugin.cpp
             ${CMAKE_SOURCE_DIR}/include/illixr/data_format/pose_prediction.hpp
             ${CMAKE_SOURCE_DIR}/include/illixr/phonebook.hpp
             ${CMAKE_SOURCE_DIR}/include/illixr/data_format/serializable_data.hpp
+            ${CMAKE_SOURCE_DIR}/include/illixr/data_format/frame.hpp
             ${CMAKE_SOURCE_DIR}/include/illixr/switchboard.hpp
             ${CMAKE_SOURCE_DIR}/include/illixr/threadloop.hpp
             ${CMAKE_SOURCE_DIR}/include/illixr/vk/display_provider.hpp
