+++ utils/serializable_data.cpp
@@ -1,8 +1,4 @@
-//
-// Created by steven on 11/5/23.
-//
-
-#include "illixr/data_format/serializable_data.hpp"
+#include "serializable_data.hpp"
 
 BOOST_CLASS_EXPORT_IMPLEMENT(ILLIXR::switchboard::event)
 BOOST_CLASS_EXPORT_IMPLEMENT(ILLIXR::data_format::compressed_frame)
